+++ src/test/java/test/TestSDSParser.java
@@ -5,6 +5,7 @@ import java.util.function.Function;
 
 import be.baur.sda.DataNode;
 import be.baur.sda.SDA;
+import be.baur.sds.DataNodeType;
 import be.baur.sds.DataType;
 import be.baur.sds.Schema;
 import be.baur.sds.parsing.SDSParser;
@@ -19,9 +20,9 @@ public final class TestSDSParser {
 		
 		/* register custom types */
 		DataType.register(IBAN.TYPE, IBAN.CONSTRUCTOR); 
-		Schema.registerDataType(IBAN.TYPE, IBANNodeType::new); 
+		DataNodeType.register(IBAN.TYPE, IBANNodeType::new); 
 		DataType.register(GMonthDay.TYPE, GMonthDay.CONSTRUCTOR);
-		Schema.registerDataType(GMonthDay.TYPE, GMonthDayNodeType::new);
+		DataNodeType.register(GMonthDay.TYPE, GMonthDayNodeType::new);
 		
 		/* test parsing SDS from files and formatting back to SDS */
 		System.out.print("contacts ");