+++ RELEASE_NOTES.md
@@ -8,12 +8,13 @@ a few renaming and re-factoring.
 - `Renamed` Type to AbstractNodeType, AnyType to AnyNodeType, DataType to DataNodeType.
 - `Renamed` CharacterType to CharacterNodeType, ComparableType to ComparableNodeType.
 - `Renamed` StringType to StringNodeType, IntegerType to IntegerNodeType, etc...
-- `Renamed` Schema methods to register and obtain data (node) type constructors.
 - `Renamed` package be.baur.sds.serialization to be.baur.sds.parsing.
 - `Removed` Validator.setTypeName() - use validate(node,type) instead.
+- `Removed` Schema methods to register and obtain data (node) type constructors.
 - `Changed` Validator.validate(node) - no longer uses a default or set type.
 - Added Validator.validate(node,type) and validateType(node,type) methods.
-- Added DataType class for registration of data types.
+- Added DataType class to register and obtain data type constructors.
+- Added DataNodeType methods to register and obtain node type constructors.
 - Added convenience method SDS.parse(File).
 - Closed issue #5 (Review type validation).
 - Closed issue #6 (Validate value against (native) SDS data type).